fix verify commit light and trusting bugs
